2.2-databases-2/orm_migrations/requirements.txt
django~=3.2.5
psycopg2-binary
django-debug-toolbar~=3.2.1